src/ch4/pywebview-notepad/requirements.txt
Flask
pywebview

